Add a generic method for polling backend status information
The volume module now implements the new method
exlybar-module-update-status. After module init, the module's
update-timer will be set to poll using the status method if any is
implemented for the specific type of module. Before module
exit the timer will be stopped if it is running.

The volume module no longer needs to manage the timer, only to
implement the status method.