Simplify font customization
To change the fonts for color codes ^f0-^f9, it now suffices to
customize exlybar-font-candidates, a list of lists of font names,
where each ordered element corresponds to ^f0, ^f1, and so on.

Prior to this change, there were five custom font variables and the
fallback lists were internal. As well, the font map was a vector of
lambdas, which is bound to be confusing.