Selectively invalidate module glyphset caches when fonts change
Each module's glyphset maps char-codes to a pixmap and a set of
metrics (glyph data), and each module maintains a cache to avoid
unnecessary rendering and loading of glyphs. Since a glyphset maps
char-codes to glyphs, any time the font for a given char-code is
switched, the glyph for that char-code should be reloaded. Prior to
this change, the cache key was the glyph position key comprising the
font hash, the pixel size, and the glyph id, meaning that when
switching back to a previously used font, the glyphs were not being
reloaded.

This changes the caches to map char-codes to glyph position
keys. Unless the value in cache matches the key for the current glyph
position, the glyph data is reloaded.

In a separate change, it may be nice to cache glyph data by the glyph
position key to prevent re-rendering.