Add a new module for displaying information about workspaces
The function variable exlybar-workspaces-generate-list-fn may be
customized to produce a structured workspaces list using specific
window manager apis. The value returned by this function is then used
to update the module status. An implementation is provided for EXWM,
though it is not necessary to have EXWM to load or use the new module.

If EXWM is present, the corresponding implementation is configured
automatically.

A similar pattern may be followed for other window managers.